Plot heatmap of merged gene expression data.

Make a heatmap of gene expression from multiple datasets.

Usage

plotExpressionHeatmap(
  fitResult,
  lambda,
  ematMerged,
  sampleMetadata,
  annoLevels,
  annoColors = NA,
  clusterTogether = FALSE,
  geneIdOrder = NA,
  className = "class",
  classLevels = NA,
  org = "org.Hs.eg",
  maxVal = 3,
  ...
)

Arguments

fitResult

glmnet object.

lambda

value of lambda to use in fitResult.

ematMerged

matrix of gene expression for genes by samples.

sampleMetadata

data.frame of sample metadata.

annoLevels

Named list used to make the annotation argument passed to pheatmap::pheatmap(). Each name must correspond to a column name in sampleMetadata, and each item in the list must be a vector of values found in that particular column.

annoColors

Passed to annotation_colors argument of pheatmap::pheatmap().

clusterTogether

logical indicating whether to cluster the samples from each dataset together or separately.

geneIdOrder

Optional character array of Entrez Gene IDs specifying the order of genes. If NA (default), the order from makeCoefDt() is used.

className

column in sampleMetadata containing values of the response variable.

classLevels

Order of classes for the column annotations.

org

Name of package for mapping Entrez Gene IDs to gene symbols, passed to data argument of annotate::lookUp().

maxVal

Maximum absolute value of scaled and centered gene expression, used to control dynamic range of color in the heatmap.

...

Additional arguments passed to pheatmap::pheatmap().

Value

A pheatmap object.